Zos3/unify size unit (#1273)
* WIP: use a new type Unit to unify sizes

A unit currently = 1 byte of either memory or storage
this will make it either for calculating usage and capacity

* correct response status

* add logs to provisiond

this was needed to help debug issue with redirections

* use correct values for reserved cap

* WiP: mem mgmt

* finalize memoy mgmt

plus multiple fixes

* fix ci

* more ci

* ci

* fix all tests

+++ pkg/container/container.go
@@ -182,7 +182,7 @@ func (c *Module) Run(ns string, data pkg.Container) (id pkg.ContainerID, err err
 		removeRunMount(),
 		withNetworkNamespace(data.Network.Namespace),
 		withMounts(data.Mounts),
-		WithMemoryLimit(data.Memory),
+		WithMemoryLimit(uint64(data.Memory)),
 		WithCPUCount(data.CPU),
 	}
